Uber for Trades/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.2
psycopg[binary,pool]==3.2.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
httpx==0.27.2
email-validator==2.2.0
